[UnifiedPDF] [iOS] No way to unlock encrypted PDF document https://bugs.webkit.org/show_bug.cgi?id=279825 rdar://136133554
Reviewed by NOBODY (OOPS!).

Currently, on iOS, with Unified PDF enabled, we do not present any UI to
unlock an encrypted PDF document.

This patch builds on bug 279807 and makes PDFPluginPassword[Field|Form]
available on the iOS family. We also make the presentment of the
password unlock annotations platform agnostic, so that there is some way
to view encrypted documents.

We will likely iterate on the existing design here because the password
unlock flow is not up to parity with PDFHVC (smaller labels, less convenient
software keyboard, etc.) but this is a first pass.

* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginAnnotation.h:
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginAnnotation.mm:
(WebKit::PDFPluginAnnotation::create):
(WebKit::PDFPluginAnnotation::~PDFPluginAnnotation):
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginPasswordField.h:
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginPasswordField.mm:
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginPasswordForm.h:
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginPasswordForm.mm:
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginTextAnnotation.h:
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginTextAnnotation.mm:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.h:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::installPDFDocument):
(WebKit::UnifiedPDFPlugin::attemptToUnlockPDF):